+++ frontend/components/Home/index.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 // import Tick from '@/app/resources/static/images/done.svg';
 // import styles from './styles.module.css';
 
-import CarparkCard from './CarparkCard';
+import CarparkCard from '../CarparkCard/CarparkCard';
 
 const Home = () => {
   return (
     <div>
-      Hello from React! jayden here test this again
+      Hello from React! jayden here check test
 
       <CarparkCard
         carpark = {{
