+++ frontend/components/Home/styles.module.css
@@ -1,105 +1,3 @@
 .home {
   color: red;
 }
-
-.card {
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-  width: 300px;
-  height: 221px;
-  
-  border: 3px solid #27348B;
-  border-radius: 7px;
-  background-color: #27348b;
-
-  margin: auto;
-}
-
-.card:hover {
-box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
-}
-
-.topHalf {
-  /* Top Half Card Container */
-  box-sizing: border-box;
-  width: 300px;
-  height: 100px;
-  text-align: center;
-  display: table-cell;
-  vertical-align: middle;
-
-  /* Title Font */
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 28px;
-  line-height: 33px;
-  color: white;
-}
-
-.bottomHalf {
-  box-sizing: border-box; 
-  width: 300px;
-  height: 121px;
-  background-color: white;
-  border-top: 3px solid #FDC300;
-  border-radius: 0px 0px 7px 7px;
-}
-
-.description {
-  /* Description Container */
-  box-sizing: border-box;
-  width: 300px;
-  height: 79px;
-  /* background-color:green; */
-
-  /* Description Font */
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 16px;
-  line-height: 19px;
-  padding: 10px 20px;
-}
-
-.directions {
-  /* Directions Container */
-  box-sizing: border-box;
-  width: 300px;
-  height:39px;
-  border-radius: 0px 0px 7px 7px;
-  /* background: red; */
-  
-  /* Directions Font */
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 12px;
-  /* line-height: 39px; */
-  padding: 0px 20px;
-  display: flex;
-  align-items: center;
-}
-
-a:link {
-  color: inherit;
-  text-decoration: none;
-  align-self: inherit;
-}
-a:visited {
-  color: inherit;
-}
-a:hover {
-  color: #1488ca;
-}
-a:active {
-  color: #FDC300;
-}
-
-i.newWindowIcon {
-  font-size: 10px;
-  color: #323232;
-}
-i.newWindowIcon:hover {
-  color: #1488ca;
-} 
-/* Ask about this */
